refactor doctor/, example/, gql/, initcmd/, runcmd/
