+++ pkg/userFlags/doctor/doctor.go
@@ -1,4 +1,4 @@
-package userflags
+package doctor
 
 import (
 	"bufio"
