+++ pkg/userFlags/doctor/vault_test.go
@@ -1,4 +1,4 @@
-package userflags
+package doctor
 
 import (
 	"encoding/json"
