+++ pkg/userFlags/example/example_test.go
@@ -1,4 +1,4 @@
-package userflags
+package example
 
 import (
 	"bytes"
