+++ pkg/userFlags/example/examples/example-auth.hk.yaml
