+++ pkg/userFlags/initcmd/init_test.go
@@ -1,4 +1,4 @@
-package userflags
+package initcmd
 
 import (
 	"bytes"
